requirements.txt
crewai
crewai-tools
pandas
plotly
python-dotenv
langchain-google-genai
google-generativeai